alteracao do user_points de NUMERIC(2) para INT.

+++ apps/backend/src/database/migrations/createUsersTable.sql
@@ -2,5 +2,5 @@ CREATE TABLE users (
     user_id UUID PRIMARY KEY,
     user_email VARCHAR() UNIQUE NOT NULL,
     user_name VARCHAR() UNIQUE NOT NULL,
-    user_points NUMERIC(2) NOT NULL DEFAULT 0,
+    user_points INT NOT NULL DEFAULT 0,
 );
